Update README.md
This site was another very complex  (for me at least) site with a mixture of CSS Flexbox and CSS Grid coded into it. I spend a lot of time really learning and understanding the fundamentals of what was going on within the code and how everything fits precisely together to make a beautiful website that can be used in the real world. I enjoy this design  creating a large "hero" image with a banner. Unlike my portfolio site where JS was implemented for the fixed navbar at a certain point when scrolling down, this site just has a static navbar with a smaller height that stays on throughout scroll. I think learning how to do the same thing several differnent ways really instills the fundamentals of front end HTML and CSS skills.